streamlit_app/requirements.txt
streamlit
st-gsheets-connection
pandas
python-docx